time series complete

+++ database/migrations/2022_01_16_140813_add_cost_to_orders_table.php
@@ -15,7 +15,7 @@ class AddCostToOrdersTable extends Migration
     {
         Schema::table('orders', function (Blueprint $table) {
             //
-            $table->integer('cost')->default(0);
+            $table->float('cost')->default(0);
         });
     }
